Ensure field states are mutually exclusive
In getting the specs to pass it exposed an issue where the `exact?`,
`conflict?` and `partial?` predicate methods weren't mutually exclusive
to each other when these methods were being overridden in subclasses.

These are now driven from a `state` value which means only one can ever
be true. Any overrides to the implementation now need to be done in the
"underscore" version of the predicate methods.