Compare superclass positions
The position comparison now wraps the statement comparison so we can
determine the correct state for superclass positions.

This is because we're unable to currently update the superclass
positions so we don't ever want the return a state of `partial`.

The `exact`.state would also be incorrect given the superclass position
won't match the suggestion position.